Use separate DeltaMpt for each Snapshot; fix wait_for_drop at program exit;  refactor storage dir configs. (#880)

+++ core/src/storage/impls/storage_db/snapshot_db_manager_sqlite.rs
@@ -12,21 +12,23 @@ pub struct SnapshotDbManagerSqlite {
 impl SnapshotDbManagerSqlite {
     const SNAPSHOT_DB_SQLITE_DIR_PREFIX: &'static str = "sqlite_";
 
-    pub fn new(snapshot_path: String) -> Self {
-        Self {
-            snapshot_path: snapshot_path + Self::SNAPSHOT_DB_SQLITE_DIR_PREFIX,
-            force_cow: false,
+    pub fn new(snapshot_path: String) -> Result<Self> {
+        let snapshot_dir = Path::new(snapshot_path.as_str());
+        if !snapshot_dir.exists() {
+            fs::create_dir_all(snapshot_dir)?;
         }
-    }
 
-    fn get_snapshot_db_path(&self, snapshot_epoch_id: &EpochId) -> String {
-        self.snapshot_path.clone() + &snapshot_epoch_id.to_hex()
+        Ok(Self {
+            snapshot_path,
+            force_cow: false,
+        })
     }
 
     fn get_merge_temp_snapshot_db_path(
         &self, old_snapshot_epoch_id: &EpochId, delta_merkle_root: &MerkleHash,
     ) -> String {
         self.snapshot_path.clone()
+            + Self::SNAPSHOT_DB_SQLITE_DIR_PREFIX
             + "merge_temp_"
             + &old_snapshot_epoch_id.to_hex()
             + &delta_merkle_root.to_hex()
@@ -36,6 +38,7 @@ impl SnapshotDbManagerSqlite {
         &self, snapshot_epoch_id: &EpochId, merkle_root: &MerkleHash,
     ) -> String {
         self.snapshot_path.clone()
+            + Self::SNAPSHOT_DB_SQLITE_DIR_PREFIX
             + "full_sync_temp_"
             + &snapshot_epoch_id.to_hex()
             + &merkle_root.to_hex()
@@ -163,48 +166,16 @@ impl SnapshotDbManagerSqlite {
 impl SnapshotDbManagerTrait for SnapshotDbManagerSqlite {
     type SnapshotDb = SnapshotDbSqlite;
 
-    fn scan_persist_state(
-        &self, snapshot_info_map: &mut HashMap<EpochId, SnapshotInfo>,
-    ) -> Result<Vec<H256>> {
-        let mut missing_snapshots = HashMap::new();
-        for (snapshot_epoch_id, _snapshot_info) in snapshot_info_map.iter() {
-            missing_snapshots.insert(
-                [
-                    StorageConfiguration::SNAPSHOT_DIR.as_bytes(),
-                    Self::SNAPSHOT_DB_SQLITE_DIR_PREFIX.as_bytes(),
-                    snapshot_epoch_id.as_ref(),
-                ]
-                .concat(),
-                snapshot_epoch_id.clone(),
-            );
-        }
+    fn get_snapshot_dir(&self) -> String { self.snapshot_path.clone() }
 
-        // Scan the snapshot dir. Remove extra files, and return the list of
-        // missing snapshots.
-        for entry in fs::read_dir(StorageConfiguration::SNAPSHOT_DIR)? {
-            let entry = entry?;
-            let path = entry.path();
-            let dir_name = path.as_path().file_name().unwrap().to_str();
-            if dir_name.is_none() {
-                error!(
-                    "Unexpected snapshot path {}, deleted.",
-                    entry.path().display()
-                );
-                fs::remove_dir_all(entry.path())?;
-                continue;
-            }
-            let dir_name = dir_name.unwrap();
-            if !missing_snapshots.contains_key(dir_name.as_bytes()) {
-                fs::remove_dir_all(entry.path())?;
-            } else {
-                missing_snapshots.remove(dir_name.as_bytes());
-            }
-        }
+    fn get_snapshot_db_name(&self, snapshot_epoch_id: &EpochId) -> String {
+        Self::SNAPSHOT_DB_SQLITE_DIR_PREFIX.to_string()
+            + &snapshot_epoch_id.to_hex()
+    }
 
-        Ok(missing_snapshots
-            .into_iter()
-            .map(|(_path_bytes, snapshot_epoch_id)| snapshot_epoch_id)
-            .collect())
+    fn get_snapshot_db_path(&self, snapshot_epoch_id: &EpochId) -> String {
+        self.snapshot_path.clone()
+            + &self.get_snapshot_db_name(snapshot_epoch_id)
     }
 
     fn new_snapshot_by_merging(
@@ -303,9 +274,13 @@ impl SnapshotDbManagerTrait for SnapshotDbManagerSqlite {
     }
 
     fn destroy_snapshot(&self, snapshot_epoch_id: &EpochId) -> Result<()> {
-        Ok(fs::remove_dir_all(
-            self.get_snapshot_db_path(snapshot_epoch_id),
-        )?)
+        if snapshot_epoch_id.ne(&NULL_EPOCH) {
+            Ok(fs::remove_dir_all(
+                self.get_snapshot_db_path(snapshot_epoch_id),
+            )?)
+        } else {
+            Ok(())
+        }
     }
 
     fn new_temp_snapshot_for_full_sync(
@@ -336,10 +311,8 @@ use crate::storage::{
         storage_db::snapshot_db_sqlite::*,
     },
     storage_db::{SnapshotDbManagerTrait, SnapshotDbTrait, SnapshotInfo},
-    StorageConfiguration,
 };
-use cfx_types::H256;
 use fs_extra::dir::CopyOptions;
 use parity_bytes::ToPretty;
 use primitives::{EpochId, MerkleHash, MERKLE_NULL_NODE, NULL_EPOCH};
-use std::{collections::HashMap, fs, process::Command};
+use std::{fs, path::Path, process::Command};
